# pet-services-api/internal/transport/http/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is an API for managing restaurants.",
        "title": "Pedido Simples API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Pedido Simples",
            "url": "http://www.pedidosimples.com.br",
            "email": "contato@pedidosimples.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/reviews": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Submeter avaliação",
                "parameters": [
                    {
                        "description": "Dados da avaliação",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/review.SubmitReviewInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Avaliação criada",
                        "schema": {
                            "$ref": "#/definitions/review.Review"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            }
        },
        "/api/v1/reviews/provider/{provider_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reviews"
                ],
                "summary": "Listar avaliações do prestador",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do prestador",
                        "name": "provider_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limite",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de avaliações",
                        "schema": {
                            "$ref": "#/definitions/http.ListReviewsResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/exceptions.ProblemDetailsOutputDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "exceptions.ProblemDetails": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "exceptions.ProblemDetailsOutputDTO": {
            "type": "object",
            "properties": {
                "problem_details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/exceptions.ProblemDetails"
                    }
                }
            }
        },
...
        "http.ListReviewsResponseDTO": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/review.Review"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "review.Review": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ownerID": {
                    "type": "string"
                },
                "providerID": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "requestID": {
                    "type": "string"
                }
            }
        },
        "review.SubmitReviewInput": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "ownerID": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "requestID": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}